Rip out uClibc internationalization (doesn't contain any codepages, just copies 'em from the host, so won't build on a host that hasn't got internationalization already installed), use gettext-stub instead of gettext, and fix udev and vim to build.

+++ sources/control-images/lfs-bootstrap/mnt/build/udev.sh
@@ -1,17 +1,20 @@
 #!/bin/sh
 
+mkdir -p /lib/udev/devices || exit 1
+if [ ! -e /lib/udev/devices/null ]
+then
+  mknod -m0666 /lib/udev/devices/null c 1 3 || exit 1
+fi
+sed -i 's/ libudev-install-move-hook//' Makefile.in &&
 install -d /lib/firmware /lib/udev/devices/pts /lib/udev/devices/shm &&
-mknod -m0666 /lib/udev/devices/null c 1 3 &&
-./configure --prefix=/usr --sysconfdir=/etc --sbindir=/sbin \
+./configure --sysconfdir=/etc --sbindir=/sbin \
   --with-rootlibdir=/lib --libexecdir=/lib/udev --disable-extras \
   --disable-introspection &&
 make -j $CPUS || exit 1
-
 if [ ! -z "$CHECK" ]
 then
   make check || exit 1
 fi
-
 make install &&
 rmdir /usr/share/doc/udev &&
 cd /mnt/packages/udev-config &&
@@ -19,5 +22,5 @@ make -j $CPUS install || exit 1
 
 if [ ! -z "$DOCS" ]
 then
-  make intall-doc
+  make install-doc
 fi
